feat: move project-error to projects feature shared errors
