+++ server/src/features/projects/errors/index.ts
@@ -1,1 +1,2 @@
 export * as ProjectCacheError from "./project-cache-page.error";
+export * as ProjectError from "./project-error";
