+++ server/src/features/projects/services/project-data-service/index.ts
@@ -1,3 +1,2 @@
-export * from "./project-error";
 export * from "./project-data-service";
 export * from "./project-data-service.type";
