+++ server/src/features/projects/services/project-data-service/project-data-service.ts
@@ -1,12 +1,12 @@
 import { CACHE } from "@/data";
 import { fail, success } from "@/utils";
+import { ProjectError } from "../../errors";
 import { IProjectCache, ProjectFilter } from "../../types";
 import { ProjectFilterUtil } from "../../utils";
 import * as FetchService from "../project-db-fetch.service";
 import { fetchProjectsData } from "./fetch-projects-data";
 import { getProjectDataKey } from "./get-project-data-key";
 import * as CacheManager from "./project-cache-manager";
-import { ProjectError } from "./project-error";
 import { ProjectPageResult, ProjectResult } from "./project-data-service.type";
 
 export async function createProjectDataService() {
@@ -47,7 +47,7 @@ export class ProjectDataService {
     return project
       ? success(project, pageResult.source)
       : fail(
-          new ProjectError({
+          new ProjectError.ProjectError({
             name: "MISSING_PROJECT_IN_PAGE_ERROR",
             message:
               "Project with specified id could not be found in the projects page.",
@@ -73,7 +73,7 @@ export class ProjectDataService {
     const pageResult = await this._cacheManager.getPage(cache, pageNumber);
     if (!pageResult.success && pageResult.error.name === "MISSING_PAGE_ERROR") {
       if (resultRecord.source === "BACKUP_CACHE") {
-        const error = new ProjectError({
+        const error = new ProjectError.ProjectError({
           name: "BACKUP_CACHE_READONLY_MODIFICATION_ERROR",
           message: "Back up cache is readonly and cannot be modified.",
         });
@@ -188,7 +188,7 @@ export class ProjectDataService {
 
     if (backupResult.success) return backupResult; //  success loading backup.
 
-    const error = new ProjectError({
+    const error = new ProjectError.ProjectError({
       name: "RESOLVE_PROJECTS_ERROR",
       message: "All fall back methods for resolving projects failed.",
     });
