+++ server/src/features/projects/services/project-data-service/project-data-service.type.ts
@@ -1,11 +1,11 @@
 import { ResultFail, ResultSuccess } from "@/types";
+import { ProjectError } from "../../errors";
 import { IProjectCache, IProjectCachePage } from "../../types";
-import { ProjectError } from "./project-error";
 
 export type Source = "DATABASE" | "JSON_CACHE" | "BACKUP_CACHE";
 export type ProjectResult =
   | ResultSuccess<IProjectCache, Source>
-  | ResultFail<ProjectError>;
+  | ResultFail<ProjectError.ProjectError>;
 export type ProjectPageResult =
   | ResultSuccess<IProjectCachePage, Source>
-  | ResultFail<ProjectError>;
+  | ResultFail<ProjectError.ProjectError>;
